api/requirements.txt
Flask
flask-restx
torch
Pillow
numpy
ultralytics